// client/src/shared/api/generated/models/image-read.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenGraph API
 * AI/ML model and dataset Web3 blockchain infrastructure server
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Image read schema
 * @export
 * @interface ImageRead
 */
export interface ImageRead {
    /**
     * File name
     * @type {string}
     * @memberof ImageRead
     */
    'file_name': string;
    /**
     * Image URL
     * @type {string}
     * @memberof ImageRead
     */
    'image_url': string;
    /**
     * Image width in pixels
     * @type {number}
     * @memberof ImageRead
     */
    'width': number;
    /**
     * Image height in pixels
     * @type {number}
     * @memberof ImageRead
     */
    'height': number;
    /**
     * Dataset ID
     * @type {number}
     * @memberof ImageRead
     */
    'dataset_id': number;
    /**
     * Image ID
     * @type {number}
     * @memberof ImageRead
     */
    'id': number;
    /**
     * Creation timestamp
     * @type {string}
     * @memberof ImageRead
     */
    'created_at': string;
}